requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
xlrd==2.0.1

